+++ internal/consts/meta.go
@@ -2,5 +2,5 @@ package consts
 
 // Meta information of the application
 
-var Version = "0.14.0-alpha.1"
-var ProtocolVersion = "0.14.0-alpha.1"
+var Version = "0.14.0-alpha.2"
+var ProtocolVersion = "0.14.0-alpha.2"