+++ internal/transport/server/pubsub/pubsub.go
@@ -1,20 +1,33 @@
 package pubsub
 
 import (
+	"context"
 	"strings"
 	"sync"
 
 	"github.com/TimeleapLabs/timeleap/internal/consts"
+	"github.com/TimeleapLabs/timeleap/internal/model"
+	"github.com/TimeleapLabs/timeleap/internal/transport/server/packet"
+	"github.com/TimeleapLabs/timeleap/internal/transport/server/websocket/queue"
+	"github.com/TimeleapLabs/timeleap/internal/utils"
 )
 
-var topics = make(map[consts.Channels][]chan []byte)
+type Subscriber struct {
+	Writer            *queue.WebSocketWriter
+	Channel           chan []byte
+	Context           *context.Context
+	ConnectionContext *context.Context
+	Unsubscribe       context.CancelFunc
+}
+
+var topics = make(map[string][]Subscriber)
 var mu sync.Mutex
 
-func getTopicsByPrefix(topic consts.Channels) map[consts.Channels][]chan []byte {
-	keys := make(map[consts.Channels][]chan []byte)
+func getTopicsByPrefix(topic string) map[string][]Subscriber {
+	keys := make(map[string][]Subscriber)
 	for key := range topics {
-		if strings.HasPrefix(string(topic), string(key)) {
-			keys[key] = make([]chan []byte, len(topics[key]))
+		if strings.HasPrefix(topic, key) {
+			keys[key] = make([]Subscriber, len(topics[key]))
 			copy(keys[key], topics[key])
 		}
 	}
@@ -22,40 +35,79 @@ func getTopicsByPrefix(topic consts.Channels) map[consts.Channels][]chan []byte 
 	return keys
 }
 
-func Publish(destinationTopic consts.Channels, operation consts.OpCode, message []byte) {
+func writeToChannel(ch chan []byte, message []byte) {
+	ch <- message
+}
+
+func PublishMessage(message []byte) {
+	msg := new(model.Message).FromBytes(message[1:])
+	Publish(msg.Topic, message)
+}
+
+func Publish(destinationTopic string, message []byte) {
 	mu.Lock()
 	defer mu.Unlock()
 
 	allSubTopics := getTopicsByPrefix(destinationTopic)
+	broadcast := append([]byte{byte(consts.OpCodeBroadcast)}, message...)
+	payload := packet.New(broadcast).Sia().Bytes()
 
 	for _, subscribers := range allSubTopics {
-		for _, ch := range subscribers {
-			go func(ch chan []byte) {
-				ch <- append([]byte{byte(operation)}, message...)
-			}(ch)
+		for _, sub := range subscribers {
+			go writeToChannel(sub.Channel, payload)
 		}
 	}
 }
 
-func Unsubscribe(topic string, ch chan []byte) {
+func Unsubscribe(topic string, writer *queue.WebSocketWriter) {
 	mu.Lock()
 	defer mu.Unlock()
 
-	for key, subscribers := range topics[consts.Channels(topic)] {
-		if subscribers == ch {
-			topics[consts.Channels(topic)] = append(topics[consts.Channels(topic)][:key], topics[consts.Channels(topic)][key+1:]...)
+	utils.Logger.
+		With("IP", writer.Conn.RemoteAddr().String()).
+		With("Topic", topic).
+		Info("Unsubscribed")
+
+	for key, sub := range topics[topic] {
+		if sub.Writer == writer {
+			topics[topic] = append(topics[topic][:key], topics[topic][key+1:]...)
+			sub.Unsubscribe()
+			close(sub.Channel)
 			break
 		}
 	}
-
-	close(ch)
 }
 
-func Subscribe(topic string) chan []byte {
+func IsSubscribed(topic string, writer *queue.WebSocketWriter) bool {
+	for _, sub := range topics[topic] {
+		if sub.Writer == writer {
+			return true
+		}
+	}
+
+	return false
+}
+
+func Subscribe(ctx context.Context, writer *queue.WebSocketWriter, topic string) (context.Context, Subscriber) {
 	mu.Lock()
 	defer mu.Unlock()
 
-	ch := make(chan []byte)
-	topics[consts.Channels(topic)] = append(topics[consts.Channels(topic)], ch)
-	return ch
+	// TODO: This is a temporary fix to prevent a panic in the tests
+	if writer.Conn != nil {
+		utils.Logger.
+			With("IP", writer.Conn.RemoteAddr().String()).
+			With("Topic", topic).
+			Info("Subscribed")
+	}
+
+	subCtx, cancel := context.WithCancel(ctx)
+
+	subscriber := Subscriber{
+		Writer:      writer,
+		Channel:     make(chan []byte),
+		Unsubscribe: cancel,
+	}
+
+	topics[topic] = append(topics[topic], subscriber)
+	return subCtx, subscriber
 }