+++ internal/transport/server/pubsub/pubsub_test.go
@@ -1,33 +1,57 @@
 package pubsub
 
 import (
+	"context"
 	"testing"
 
+	"github.com/TimeleapLabs/timeleap/internal/config"
 	"github.com/TimeleapLabs/timeleap/internal/consts"
+	"github.com/TimeleapLabs/timeleap/internal/crypto"
+	"github.com/TimeleapLabs/timeleap/internal/transport/server/websocket/queue"
+	"github.com/TimeleapLabs/timeleap/internal/utils"
 )
 
+func TestMain(m *testing.M) {
+	utils.SetupLogger("info")
+	config.App.System.AllowGenerateSecrets = true
+
+	crypto.InitMachineIdentity(
+		crypto.WithEd25519Identity(),
+	)
+
+	m.Run()
+}
+
 func TestSubscribeDirectly(t *testing.T) {
-	sub := Subscribe("a.b.c")
+	ctx := context.Background()
+	writer := queue.NewWebSocketWriter(nil, 1)
+	_, sub := Subscribe(ctx, writer, "a.b.c")
 
-	Publish("a.b.c", consts.OpCodeAttestation, []byte("Hello, world!"))
+	Publish("a.b.c", []byte("Hello, world!"))
 
-	msg := <-sub
-	if string(msg[1:]) != "Hello, world!" {
-		t.Fatalf("Received unexpected message: %s", msg)
+	msg := <-sub.Channel
+	opcode := msg[0]
+	if opcode != byte(consts.OpCodeBroadcast) {
+		t.Fatalf("Received unexpected opcode: %d", opcode)
+	}
+
+	str := string(msg[1 : len(msg)-96])
+	if str != "Hello, world!" {
+		t.Fatalf("Received unexpected message: '%s'", str)
 	}
 }
 
 func TestGetTopicsByPrefix(t *testing.T) {
 	const (
-		first  consts.Channels = "a.b.c"
-		second consts.Channels = "a"
-		third  consts.Channels = "b.c.d"
+		first  string = "a.b.c"
+		second string = "a"
+		third  string = "b.c.d"
 	)
 
-	topics = map[consts.Channels][]chan []byte{
-		first:  make([]chan []byte, 0),
-		second: make([]chan []byte, 0),
-		third:  make([]chan []byte, 0),
+	topics = map[string][]Subscriber{
+		first:  {},
+		second: {},
+		third:  {},
 	}
 
 	trimmedTopics := getTopicsByPrefix(first)
@@ -40,17 +64,24 @@ func TestGetTopicsByPrefix(t *testing.T) {
 }
 
 func TestSubscribeWithPrefix(t *testing.T) {
-	sub := Subscribe("a")
+	ctx := context.Background()
+	writer := queue.NewWebSocketWriter(nil, 1)
+	_, sub := Subscribe(ctx, writer, "a")
 
-	Publish("a.b.c", consts.OpCodeAttestation, []byte("Hello, world!"))
+	Publish("a.b.c", []byte("Hello, world!"))
 
-	msg := <-sub
+	msg := <-sub.Channel
+	opcode := msg[0]
+	if opcode != byte(consts.OpCodeBroadcast) {
+		t.Fatalf("Received unexpected opcode: %d", opcode)
+	}
 
-	if string(msg[1:]) != "Hello, world!" {
-		t.Fatalf("Received unexpected message: %s", msg)
+	str := string(msg[1 : len(msg)-96])
+	if str != "Hello, world!" {
+		t.Fatalf("Received unexpected message: '%s'", str)
 	}
 }
 
 func TestPublishWithoutSubscriber(_ *testing.T) {
-	Publish("a.b.c", consts.OpCodeAttestation, []byte("Hello, world!"))
+	Publish("a.b.c", []byte("Hello, world!"))
 }