+++ internal/transport/server/websocket/websocket.go
@@ -7,6 +7,7 @@ import (
 	"log/slog"
 	"net/http"
 
+	sia "github.com/TimeleapLabs/go-sia/v2/pkg"
 	"github.com/TimeleapLabs/timeleap/internal/consts"
 	"github.com/TimeleapLabs/timeleap/internal/transport/server/packet"
 	"github.com/TimeleapLabs/timeleap/internal/transport/server/pubsub"
@@ -100,24 +101,19 @@ func multiplexer(w http.ResponseWriter, r *http.Request) {
 			continue
 		}
 
+		// TODO: This needs to be refactored
 		switch consts.OpCode(p.Message[0]) {
-		case consts.OpCodeAttestation:
-			result, err := handler.AttestationRecord(p.Message[1:], p.Signature, p.Signer)
-			if err != nil {
-				writer.SendError(consts.OpCodeError, err)
-				continue
+		case consts.OpCodeUnSubscribe:
+			topic := sia.NewFromBytes(payload[1:]).ReadString16()
+			go pubsub.Unsubscribe(topic, writer)
+		case consts.OpCodeSubscribe:
+			topic := sia.NewFromBytes(payload[1:]).ReadString16()
+			if !pubsub.IsSubscribed(topic, writer) {
+				subCtx, sub := pubsub.Subscribe(ctx, writer, topic)
+				go handler.BroadcastManager(ctx, subCtx, topic, sub)
 			}
-
-			pubsub.Publish(consts.ChannelAttestation, consts.OpCodeAttestation, result)
-			writer.SendMessage(consts.OpCodeFeedback, "signature.accepted")
-		case consts.OpCodeRegisterConsumer:
-			utils.Logger.
-				With("IP", conn.RemoteAddr().String()).
-				With("Channel", string(payload[1:])).
-				Info("New Consumer registered")
-
-			topic := string(payload[1:])
-			go handler.BroadcastListener(ctx, conn, topic, pubsub.Subscribe(topic))
+		case consts.OpCodeMessage:
+			go pubsub.PublishMessage(payload)
 		case consts.OpCodeRegisterWorker:
 			isWorker = true
 			go handler.RegisterWorker(ctx, conn, payload[1:])